+++ packages/in_app_purchase/in_app_purchase_storekit/test/in_app_purchase_storekit_platform_test.dart
@@ -489,6 +489,38 @@ void main() {
       expect(
           fakeStoreKitPlatform.finishedTransactions.first.payment.quantity, 5);
     });
+
+    test(
+        'buying non consumable with discount, should get purchase objects in the purchase update callback',
+        () async {
+      final List<PurchaseDetails> details = <PurchaseDetails>[];
+      final Completer<List<PurchaseDetails>> completer =
+          Completer<List<PurchaseDetails>>();
+      final Stream<List<PurchaseDetails>> stream =
+          iapStoreKitPlatform.purchaseStream;
+
+      late StreamSubscription<List<PurchaseDetails>> subscription;
+      subscription = stream.listen((List<PurchaseDetails> purchaseDetailsList) {
+        details.addAll(purchaseDetailsList);
+        if (purchaseDetailsList.first.status == PurchaseStatus.purchased) {
+          completer.complete(details);
+          subscription.cancel();
+        }
+      });
+      final AppStorePurchaseParam purchaseParam = AppStorePurchaseParam(
+        productDetails:
+            AppStoreProductDetails.fromSKProduct(dummyProductWrapper),
+        applicationUserName: 'userWithDiscount',
+        discount: dummyPaymentDiscountWrapper,
+      );
+      await iapStoreKitPlatform.buyNonConsumable(purchaseParam: purchaseParam);
+
+      final List<PurchaseDetails> result = await completer.future;
+      expect(result.length, 2);
+      expect(result.first.productID, dummyProductWrapper.productIdentifier);
+      expect(fakeStoreKitPlatform.discountReceived,
+          dummyPaymentDiscountWrapper.toMap());
+    });
   });
 
   group('complete purchase', () {