+++ packages/espresso/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0+5
+
+* Updates android gradle plugin to 7.3.1.
+
 ## 0.2.0+4
 
 * Updates minimum Flutter version to 2.10.