+++ packages/file_selector/file_selector/CHANGELOG.md
@@ -1,5 +1,7 @@
-## NEXT
+## 0.9.2+2
 
+* Improves API docs and examples.
+* Changes XTypeGroup initialization from final to const.
 * Updates minimum Flutter version to 2.10.
 
 ## 0.9.2