+++ packages/shared_preferences/shared_preferences_android/CHANGELOG.md
@@ -1,5 +1,6 @@
-## NEXT
+## 2.0.14
 
+* Fixes typo in `SharedPreferencesAndroid` docs.
 * Updates code for `no_leading_underscores_for_local_identifiers` lint.
 
 ## 2.0.13